Refactor data display into DataMap
